+++ packages/frontend/src/components/workflow-run-debugger/components/panels/StepInspectorPanel.tsx
@@ -26,7 +26,7 @@ export const StepInspectorPanel = () => {
           gap: 2,
         }}
       >
-        <Typography variant="subtitle1" fontWeight={600}>
+        <Typography variant="subtitle2" fontWeight={600}>
           {t("label.step_inspector")}
         </Typography>
         <Box
