+++ packages/frontend/src/components/workflow-run-debugger/components/panels/step-trace-panel/StepTraceHeader.tsx
@@ -8,20 +8,13 @@ import { Box, Typography } from "@mui/material";
 
 import { useTranslate } from "@/hooks/useTranslate";
 
-type StepTraceHeaderProps = {
-  status?: string | null;
-};
-
-export const StepTraceHeader = ({ status }: StepTraceHeaderProps) => {
+export const StepTraceHeader = () => {
   const { t } = useTranslate();
 
   return (
     <Box display="flex" alignItems="center" justifyContent="space-between">
-      <Typography variant="subtitle1" fontWeight={600}>
+      <Typography variant="subtitle2" fontWeight={600}>
         {t("label.step_trace.title")}
-      </Typography>
-      <Typography variant="body2" color="text.secondary">
-        {status ?? "—"}
       </Typography>
     </Box>
   );
