+++ packages/frontend/src/layout/theme/customizations/inputs.tsx
@@ -302,7 +302,7 @@ export const inputsCustomizations: Components<Theme> = {
   MuiToggleButton: {
     styleOverrides: {
       root: ({ theme }) => ({
-        padding: "12px 16px",
+        padding: "6px 12px",
         textTransform: "none",
         borderRadius: "10px",
         fontWeight: 500,
