+++ packages/frontend/src/routes/routeConfig.tsx
@@ -30,7 +30,7 @@ import { Subscribers } from "@/components/subscribers";
 import { Translations } from "@/components/translations";
 import { Users } from "@/components/users";
 import { WorkflowEditor } from "@/components/visual-editor/v4";
-import { WorkflowRunDebugger } from "@/components/workflow-run-debugger";
+import { WorkflowRunDebuggerPage } from "@/components/workflow-run-debugger";
 import { WorkflowRuns } from "@/components/workflow-runs";
 import { LayoutProps } from "@/layout";
 import { EntityType } from "@/services/types";
@@ -89,8 +89,8 @@ export const routes: RouteObjectItem[] = [
     },
   },
   {
-    path: "/workflow/:workflowId/runs/:InitiatorId",
-    Component: WorkflowRunDebugger,
+    path: "/workflow/:workflowId/runs/:initiatorId",
+    Component: WorkflowRunDebuggerPage,
     handle: {
       requiredPermissions: [[EntityType.WORKFLOW_RUN, PermissionAction.READ]],
     },
